fix(api)!: make authorID a path instead of a body
I had not fully tested this and so it never really worked properly when I made it a body.
Meaning the route did not work at all.

Now I made it a path and not a query because I feel it'd make more sense for it to be `/{authorID}?id=...` instead of `/{id}?authorID=...` as you are deleting an a user's preview ID.
So rather it act like getAll's `GET /{authorID}`and make authorID the path.